requirements.txt
census==0.8.19
contextily==1.2.0
geopandas==0.10.2
matplotlib==3.5.1
pandas==1.3.2
scikit_learn==1.2.2
scipy==1.7.3
seaborn==0.11.2
statsmodels==0.13.0
us==2.0.2
python-dotenv==1.0.0
